Finished

+++ rsa_bali/src/parsing/parsing.h
@@ -1,7 +1,7 @@
 #ifndef __PARSING_H
 #define __PARSING_H
 
-void encode(unsigned char *outmsg, unsigned long *outblock, unsigned char blocksiz);
-unsigned char decode(unsigned long *inblock, unsigned char *inmsg);
+void encode(unsigned char *outmsg, unsigned long long *outblock, unsigned char blocksiz);
+unsigned char decode(unsigned long long *inblock, unsigned char *inmsg);
 
 #endif /* __PARSING_H */
